feat(store): Change candidate store to RaceData
Create new store to fetch all candidates, change original candidate store to raceData, create candidates route

Progress on #19
