+++ client/store/candidates.js
@@ -3,58 +3,33 @@ import axios from 'axios'
 /**
  * ACTION TYPES
  */
-const ADD_CANDIDATE = 'ADD_CANDIDATE'
-const EDIT_CANDIDATE = 'EDIT_CANDIDATE'
+const GET_CANDIDATES = 'GET_CANDIDATES'
 
 /**
  * INITIAL STATE
  */
-const defaultCandidatesState = []
+const defaultCandidateState = []
 
 /**
  * ACTION CREATORS
  */
-export const createdCandidate = candidate => ({
-  type: ADD_CANDIDATE,
-  candidate
-})
-
-export const editedCandidate = (candidateId, candidate) => ({
-  type: EDIT_CANDIDATE,
-  candidateId,
-  candidate
+export const gotCandidates = candidates => ({
+  type: GET_CANDIDATES,
+  candidates
 })
 
 /**
  * THUNK CREATORS
  */
-export const postCandidate = (raceId, candidate) => {
+export const getCandidates = () => {
   return async dispatch => {
     try {
-      const response = await axios.post(
-        `api/races/${raceId}/candidates`,
-        candidate
-      )
-      const newCandidate = response.data
-      const action = createdCandidate(newCandidate)
+      const response = await axios.get('/api/candidates')
+      const candidates = response.data
+      const action = gotCandidates(candidates)
       dispatch(action)
     } catch (err) {
-      console.error(err)
-    }
-  }
-}
-
-export const putCandidate = (raceId, candidateId, candidate) => {
-  return async dispatch => {
-    try {
-      const response = await axios.put(
-        `api/races/${raceId}/candidates/${candidateId}`,
-        candidate
-      )
-      const updatedCandidate = response.data
-      const action = editedCandidate(candidateId, updatedCandidate)
-      dispatch(action)
-    } catch (err) {
+      // to add toastr
       console.error(err)
     }
   }
@@ -63,19 +38,10 @@ export const putCandidate = (raceId, candidateId, candidate) => {
 /**
  * REDUCER
  */
-export default function(state = defaultCandidatesState, action) {
+export default function(state = defaultCandidateState, action) {
   switch (action.type) {
-    case ADD_CANDIDATE: {
-      return [...state, action.candidate]
-    }
-    case EDIT_CANDIDATE: {
-      const editedCandidates = state.candidates.map(candidate => {
-        if (candidate.id === action.candidateId) {
-          return action.candidate
-        }
-        return candidate
-      })
-      return [...state, editedCandidates]
+    case GET_CANDIDATES: {
+      return action.candidates
     }
     default:
       return state
