+++ client/store/index.js
@@ -4,9 +4,10 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 import races from './races'
+import raceData from './raceData'
 import candidates from './candidates'
 
-const reducer = combineReducers({user, races, candidates})
+const reducer = combineReducers({user, races, raceData, candidates})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
