+++ server/api/index.js
@@ -4,6 +4,7 @@ module.exports = router
 router.use('/users', require('./users'))
 router.use('/orders', require('./orders'))
 router.use('/races', require('./races'))
+router.use('/candidates', require('./candidates'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
